requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
git+https://github.com/Rapptz/asqlite
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
braceexpand==0.1.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colour==0.1.5
git+https://github.com/Rapptz/discord.py
idna==3.2
import-expression==1.1.4
jishaku==2.1.0
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3